Refactor `StartProcess` to new way (#40)
Whew, this PR originally started with the goal to remove the restorer/subsriber from the process directory in order to make the logic flow more easy to follow and rewrite the flow so that NodeController (or something like it) is a central place of control. I ended up doing a bit more than that as I came across places for optimization and cleanup. Gonna try to write down all the changes here, but I might have missed a couple things:

Create a node.Component[T] type that can be used to force modules used by the node controller to be able to restore certain state and certain transitions. SupportedTransitionTypes() is not actually used anywhere (yet -- I'm considering being a bit stricter about typing at some point), so I'm considering removing it.
Start creating a CtrlServer. This will eventually completely replace the apiServer we have going, and as a result we can stop needing this server having to implement Route and simply serve these routes normally.
Start creating a controller2. This needs a new name. This is the actual backing for CtrlServer, which is just a http server wrapper around this. The goal here is that it will eventually do everything BaseNodeController does, but without exporting these functions, because everything that needs to be called out to will happen internally.
After doing that, I realized that there is a bit of redundancy in the node transitions and the different process states.

Specifically, process can appear in the node state as either Starting or Running, but these are always bound to get out of sync. For e.g. you could mark something as Running but then the process crashes. Or it's possible to be in Running when it says Starting. I replaced both of these with Started. This indicates that we would like to be in a running state, but does not distinguish if we have achieved that (since that distinguishment can be incorrect anyways). I believe this gives us the exact same functionality and is easier to reason about.
After doing this I realized the Stopped state wasn't actually used anywhere -- existence indicated a Started process and non-existence indicated nothing was there. So I removed the states completely to directly reflect this. I had to add a migration for this.
Stop enriching StartProcessTransition. I have a gut feeling that we don't actually need to be enriching any transitions and should restructure our code to not do this. It's really hard to reason about the state at any point with these Enrich functions. I will try to remove each of these and then fully remove that method from the interface. In this case, after I did 1 + 2, we didn't need the extra running step which adds we  Ext_Driver_ID, because the node doesn't do anything with it. Conceptually, this is something that should be managed by just the process manager anyways. The one exception to this is that it's possible for the node to crash or something, leaving a process with Ext_Driver_ID running, but the opposite is possible too, the node thinks some Ext_Driver_ID is running but it's not, or there's many processes we've spun up and the node doesn't know about all of them. Instead, what we should do is to have drivers tag processes with the ID that we generate in the StartProcessTransition, and then have the drivers be able to check if they are still there, and launch them or kill them if needed. I will work on doing this in a follow-up -- we should do this with PDS so habitat doesn't try to launch it every time we re-run even though there's a process already running
---------

Co-authored-by: Arushi Bandi <arushibandi@users.noreply.github.com>