Make process manager more robust (#42)
The process manager was a bit broken because the running processes map got reset every time habitat got restarted, but it's possible that there were still processes running in docker. So for example,
ListProcesses and GetProcess wouldn't report those processes so the CLI tool was broken.

In general, the actual state of processes can drift from both what's in the node state and whatever is in our in-memory state in the process manager. For the latter, it's best to remove this state and pass all API calls through to the process driver which is backed by docker / something that can check the actual state in the moment. That's what this PR does.

For the former issue, we need to be careful about when we commit transitions or not. This is the code in controller2.go and we should be careful about writing a transition if the corresponding action failed. I handled a couple edge cases here -- the process is already started or already stopped -- there might be a couple more but I think it's good enough for now

Also make type Driver string so we don't have "docker" everywhere

---------

Co-authored-by: Arushi Bandi <arushibandi@users.noreply.github.com>