controller cleanup (#58)
Delete the old BaseNodeController type in favor of Controller2.
Also delete the PDSClient since we can either pass through directly to PDS from the front end or call via an xrpc.Client.
Remove certificates from users -- we should just rely on DID. We don't use these anywhere.
^ As a result, we can remove the auth middleware. It's unused. We should rely on PDS did.
Remove the Enrich() functions on transitions. Only one of these was being used and it could easily be moved to the Gen function for it. We should have stricter typing around the Gen functions and make all the transition types private but leaving this as a follow up.
Test Plan:

I tested from the frontend making a new user and logging in my existing users and it works.
I also will fix up or add unit tests.