[Integration] Add OAuth integration test with real PDS (#23)
use https://golang.testcontainers.org/ to spin up real instances of
frontend, privi, and a pds to actually validate e2e oauth flow.

published a test web did at
https://sashankg.github.io/.well-known/did.json. we can move it later

tried to make it modular to support other integration tests as well.
probably more work needed to clean it up further

we should NOT add this as a PR blocker imo. it already takes ~4 min and
it should take longer as we add more tests. rn it's manual run only but
maybe we can run it nightly. once we have deploys, it should be a
requirement for that.

example successful run:
https://github.com/habitat-network/habitat/actions/runs/20085439732/job/57621735520